requirements.txt
mysqlclient
Flask==3.0.0
keras-tcn==3.5.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
SQLAlchemy==2.0.23
gunicorn==21.2.0
python-dotenv==1.0.0
tensorflow==2.15.0
tf-keras==2.15.0